+++ pyoffers/models/country.py
@@ -9,3 +9,8 @@ class Country(Model):
 class CountryManager(ModelManager):
     model = Country
     name = 'countries'
+
+    def init_instance(self, data):
+        if 'id' not in data:
+            data = list(data.values())[0]
+        return super().init_instance(data)